Move methods for inspecting mappings to a separate file
